requirements.txt
python-dotenv==0.18.0
numpy==1.21.0
opencv-python==4.5.3.56
requests==2.26.0
Pillow==8.3.1
schedule==1.1.0